+++ src/main/java/com/wanted/feed/common/exception/GlobalControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.wanted.feed.exception;
+package com.wanted.feed.common.exception;
 
 import java.util.StringJoiner;
 import lombok.extern.slf4j.Slf4j;
