--- src/main/java/com/wanted/feed/exception/client/SnsContentIdNotNullException.java
@@ -1,7 +1,0 @@
-package com.wanted.feed.exception.client;
-
-import com.wanted.feed.exception.WantedException;
-
-public class SnsContentIdNotNullException extends WantedException {
-
-}
