--- src/main/java/com/wanted/feed/exception/client/SnsNotSupportException.java
@@ -1,7 +1,0 @@
-package com.wanted.feed.exception.client;
-
-import com.wanted.feed.exception.WantedException;
-
-public class SnsNotSupportException extends WantedException {
-
-}
