--- src/main/java/com/wanted/feed/exception/feed/like/LikeFeedIdNotNullException.java
@@ -1,6 +1,0 @@
-package com.wanted.feed.exception.feed.like;
-
-import com.wanted.feed.exception.WantedException;
-
-public class LikeFeedIdNotNullException extends WantedException {
-}
