+++ src/main/java/com/wanted/feed/feed/domain/vo/FeedId.java
@@ -1,6 +1,6 @@
 package com.wanted.feed.feed.domain.vo;
 
-import com.wanted.feed.exception.feed.like.LikeFeedIdNotNullException;
+import com.wanted.feed.feed.exception.like.LikeFeedIdNotNullException;
 import jakarta.persistence.Embeddable;
 import java.util.Objects;
 import lombok.AccessLevel;
