+++ src/main/java/com/wanted/feed/feed/exception/FeedNotFoundException.java
@@ -1,6 +1,6 @@
 package com.wanted.feed.feed.exception;
 
-import com.wanted.feed.exception.WantedException;
+import com.wanted.feed.common.exception.WantedException;
 
 public class FeedNotFoundException extends WantedException {
 }
