+++ src/main/java/com/wanted/feed/feed/exception/like/LikeUserIdNotNullException.java
@@ -1,0 +1,6 @@
+package com.wanted.feed.feed.exception.like;
+
+import com.wanted.feed.common.exception.WantedException;
+
+public class LikeUserIdNotNullException extends WantedException {
+}
