+++ src/main/java/com/wanted/feed/feed/exception/share/ShareUserIdNotNullException.java
@@ -1,6 +1,6 @@
 package com.wanted.feed.feed.exception.share;
 
-import com.wanted.feed.exception.WantedException;
+import com.wanted.feed.common.exception.WantedException;
 
 public class ShareUserIdNotNullException extends WantedException {
