+++ src/main/java/com/wanted/feed/feign/exception/SnsContentIdNotNullException.java
@@ -1,0 +1,7 @@
+package com.wanted.feed.feign.exception;
+
+import com.wanted.feed.common.exception.WantedException;
+
+public class SnsContentIdNotNullException extends WantedException {
+
+}
