+++ src/main/java/com/wanted/feed/user/exception/DuplicateUserException.java
@@ -1,6 +1,6 @@
 package com.wanted.feed.user.exception;
 
-import com.wanted.feed.exception.WantedException;
+import com.wanted.feed.common.exception.WantedException;
 
 public class DuplicateUserException extends WantedException {
 }
