+++ src/test/java/com/wanted/feed/feed/service/FeedShareServiceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.wanted.feed.exception.client.SnsShareFeedFailException;
+import com.wanted.feed.feign.exception.SnsShareFeedFailException;
 import com.wanted.feed.feed.domain.Feed;
 import com.wanted.feed.feed.domain.FeedRepository;
 import com.wanted.feed.feed.domain.Share;
